+++ app/products.py
@@ -1,5 +1,6 @@
 import app.connect as c
 from app import app
+from flask_login import login_required, current_user
 # Class for creating and editing product information
 class ProductAct:
     def __init__(self, product_id=None):
@@ -27,7 +28,7 @@ class ProductAct:
             # (HERE) Checking for the presence of the same product
             self.prod_exist = self.conn.execute('''SELECT p.id AS id, p.name AS name, weight, weightType, price, s.name AS shop
                                          FROM products p JOIN shops s
-                                         ON p.shop_id = s.id 
+                                         ON p.shop_id = s.id
                                          WHERE p.name=? AND s.name=?''',
                                         (self.data['name'], self.data['shop'])).fetchall()
             try:
@@ -59,6 +60,12 @@ class ProductAct:
         # Inserting product to the db
         self.conn.execute('INSERT INTO products (name, weight, weightType, price, shop_id) VALUES (?, ?, ?, ?, ?)',
                     (self.data['name'], self.data['weight'], self.data['weight-type'], self.data['price'], shop_id))
+
+        prod_id = self.conn.execute('''SELECT id FROM products 
+                                        WHERE name=? AND weight=? AND weightType=? AND price=? AND shop_id=?''',
+                                        (self.data['name'], self.data['weight'], self.data['weight-type'], self.data['price'], shop_id)).fetchone()
+        self.conn.execute('INSERT INTO usersProducts (user_id, prod_id) VALUES (?,?)',
+                            (current_user.id, prod_id[0]))
         self.conn.commit()
         self.conn.close()
         return True
@@ -76,7 +83,9 @@ def get_product(product_id):
     product = conn.execute('''SELECT p.id AS id, p.name AS name, weight, weightType, price, s.name AS shop
                               FROM products p JOIN shops s
                               ON p.shop_id = s.id
-                              WHERE p.id = ?''', (product_id,)).fetchone()
+                              JOIN usersProducts u
+                              ON u.prod_id = p.id
+                              WHERE p.id = ? AND u.user_id=?''', (product_id, current_user.id)).fetchone()
     conn.close()
     if product is None:
         c.abort(404)
@@ -84,26 +93,43 @@ def get_product(product_id):
 # List of products page (open)
 @app.route('/products')
 def products():
-    conn = c.get_db_connection()
-    products = conn.execute('''SELECT p.id AS id, p.name AS name, weight, weightType, price, s.name AS shop
-                               FROM products p JOIN shops s
-                               ON p.shop_id = s.id''').fetchall()
-    conn.close()
+    products = None
+    if current_user.is_authenticated:
+        conn = c.get_db_connection()
+        products = conn.execute('''SELECT p.id AS id, p.name AS name, weight, weightType, price, s.name AS shop
+                                FROM products p JOIN shops s
+                                ON p.shop_id = s.id
+                                JOIN usersProducts u
+                                ON u.prod_id = p.id
+                                WHERE u.user_id=?''', (current_user.id,)).fetchall()
+        conn.commit()
+        conn.close()
     return c.render_template('products.html', products = products)
+
 @app.route('/product<int:product_id>')
+@login_required
 def product(product_id):
-    product = get_product(product_id)
     conn = c.get_db_connection()
-    recipes = conn.execute('''SELECT r.id, r.name, rt.type, i.prod_id, i.weight
-                              FROM recipes r JOIN recipeTypes rt
-                              ON r.type_id = rt.id
-                              JOIN ingredients i
-                              ON r.id = i.rec_id
-                              WHERE i.prod_id = ?''', (product_id,)).fetchall()
-    conn.close()
-    return c.render_template('product.html', product=product, recipes=recipes)
+    prod_id_exist = conn.execute('SELECT id FROM usersProducts WHERE user_id=? AND prod_id=?', (current_user.id, product_id)).fetchone()
+    if not prod_id_exist:
+        conn.commit()
+        conn.close()
+        return c.redirect(c.url_for('products'))
+    else:
+        product = get_product(product_id)
+        recipes = conn.execute('''SELECT r.id, r.name, rt.type, i.prod_id, i.weight
+                                FROM recipes r JOIN recipeTypes rt
+                                ON r.type_id = rt.id
+                                JOIN ingredients i
+                                ON r.id = i.rec_id
+                                JOIN usersRecipes u
+                                ON u.rec_id = r.id
+                                WHERE i.prod_id = ? AND u.user_id=?''', (product_id, current_user.id)).fetchall()
+        conn.close()
+        return c.render_template('product.html', product=product, recipes=recipes)
 # Create new product page (open, form handling)
 @app.route('/create_product', methods=('GET', 'POST'))
+@login_required
 def create_p():
     productAct = ProductAct()
     if productAct.act(productAct.create):
@@ -111,18 +137,33 @@ def create_p():
     return c.render_template('create_p.html', shops=productAct.shops, data=productAct.data, prod_exist=productAct.prod_exist)
 # Edit product page (open, form handling)
 @app.route('/<int:product_id>_product_edit', methods=('GET', 'POST'))
+@login_required
 def edit_p(product_id):
-    productAct = ProductAct(product_id)
-    if productAct.act(productAct.edit):
+    conn = c.get_db_connection()
+    prod_id_exist = conn.execute('SELECT id FROM usersProducts WHERE user_id=? AND prod_id=?', (current_user.id, product_id)).fetchone()
+    if not prod_id_exist:
+        conn.commit()
+        conn.close()
         return c.redirect(c.url_for('products'))
-    return c.render_template('edit_p.html', shops=productAct.shops, product=productAct.data)
+    else:
+        productAct = ProductAct(product_id)
+        if productAct.act(productAct.edit):
+            return c.redirect(c.url_for('products'))
+        return c.render_template('edit_p.html', shops=productAct.shops, product=productAct.data)
 # Delete product
 @app.route('/<int:product_id>/delete_product', methods=('POST',))
+@login_required
 def delete_p(product_id):
-    product = get_product(product_id)
     conn = c.get_db_connection()
-    conn.execute('DELETE FROM products WHERE id = ?', (product_id,))
-    conn.commit()
-    conn.close()
-    c.flash('"{}" was successfully deleted!'.format(product['name']))
-    return c.redirect(c.url_for('products'))
+    prod_id_exist = conn.execute('SELECT id FROM usersProducts WHERE user_id=? AND prod_id=?', (current_user.id, product_id)).fetchone()
+    if not prod_id_exist:
+        conn.commit()
+        conn.close()
+        return c.redirect(c.url_for('products'))
+    else:
+        product = get_product(product_id)
+        conn.execute('DELETE FROM products WHERE id = ?', (product_id,))
+        conn.commit()
+        conn.close()
+        c.flash('"{}" was successfully deleted!'.format(product['name']))
+        return c.redirect(c.url_for('products'))
